add the clientlayout and clietnprovider in components
